backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.41
psycopg2-binary==2.9.9
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
websockets==15.0.1
python-dotenv==1.1.0
alembic==1.13.1
pydantic==2.11.4
pydantic-settings==2.9.1
httpx==0.28.1
python-socketio==5.11.1
email-validator==2.2.0
PyJWT==2.8.0
requests==2.31.0 